Move Alternatives section to Appendix
This moves the informative guidance text for alternative setups to
an appendix.  The reason for this is that this text is not required
for the ANAME protocol but provides useful information for providers
that use a non-traditional DNS architecture.

This also rewrites the text in the Alternatives section to include
guidance on how to deal with zone transfers scalability and tailored
responses (based on GeoIP for example).